% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hapi-package.R
\docType{package}
\name{Hapi-package}
\alias{Hapi-package}
\alias{Hapi}
\title{Hapi is a novel easy-to-use package that only requires 3 to 5 gametes 
to reconstruct accurate and high-resolution haplotypes of an individual. 
The gamete genotype data may be generated from various platforms including 
genotyping arrays and next generation sequencing even with low-coverage. 
Hapi simply takes genotype data of known hetSNPs in single gamete cells 
as input and report the high-resolution haplotypes as well as confidence 
level of each phased hetSNPs. The package also includes a module allowing 
downstream analyses and visualization of crossovers in the gametes.}
\description{
Hapi is a novel easy-to-use package that only requires 3 to 5 gametes 
to reconstruct accurate and high-resolution haplotypes of an individual. 
The gamete genotype data may be generated from various platforms including 
genotyping arrays and next generation sequencing even with low-coverage. 
Hapi simply takes genotype data of known hetSNPs in single gamete cells 
as input and report the high-resolution haplotypes as well as confidence 
level of each phased hetSNPs. The package also includes a module allowing 
downstream analyses and visualization of crossovers in the gametes.
}
