% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossover.R
\name{hapiCVDistance}
\alias{hapiCVDistance}
\title{Histogram of crossover distance}
\usage{
hapiCVDistance(cv)
}
\arguments{
\item{cv}{a dataframe of crossover information}
}
\value{
a histogram
}
\description{
Histogram of crossover distance
}
\examples{
data(crossover)
hapiCVDistance(cv=crossover)
}
\author{
Ruidong Li
}
