% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossover.R
\name{hapiCVResolution}
\alias{hapiCVResolution}
\title{Histogram of crossover resolution}
\usage{
hapiCVResolution(cv)
}
\arguments{
\item{cv}{a dataframe of crossover information}
}
\value{
a histogram
}
\description{
Histogram of crossover resolution
}
\examples{
data(crossover)
hapiCVResolution(cv=crossover)
}
\author{
Ruidong Li
}
