% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft.R
\name{hapiPhase}
\alias{hapiPhase}
\title{Phase draft haplotypes by majority voting}
\usage{
hapiPhase(gmt)
}
\arguments{
\item{gmt}{a dataframe of imputed genotype data of gamete cells}
}
\value{
a dataframe of inferred draft haplotypes
}
\description{
Phase draft haplotypes by majority voting
}
\examples{
ref <- rep(0,500)
alt <- rep(1,500)
imputedFrame <- data.frame(gmt1=ref, gmt2=alt, gmt3=ref, 
gmt4=ref, gmt5=c(alt[1:250], ref[251:500]),
stringsAsFactors = FALSE)
draftHap <- hapiPhase(gmt=imputedFrame)
}
\author{
Ruidong Li
}
